adding naive version 10% cublas

+++ include/gemm_kernels.h
@@ -1,6 +1,6 @@
 #ifndef GEMM_KERNELS_H
 #define GEMM_KERNELS_H
 
-void dgemm_cuda(double* A, double* B, double* C, int M, int N, int K);
+void dgemm_cuda(float* A, float* B, float* C, int M, int N, int K);
 
 #endif
