+++ src/dgemm.cu
@@ -1,14 +1,14 @@
 #include <cuda_runtime.h>
 #include "gemm_kernels.h"
 
-__global__ void naive_gemm(double* A, double* B, double* C, int M, int K, int N){
+__global__ void naive_gemm(float* A, float* B, float* C, int M, int K, int N){
 
 	int i = threadIdx.y + blockIdx.y * blockDim.y;
 	int j = threadIdx.x + blockIdx.x * blockDim.x;
 
 	if (i < M && j < N){
 
-		double tmp = 0.0;
+		float tmp = 0.0;
 		for (int k = 0; k<K; k++){
 			tmp += A[i*K +k] * B[k*N+ j];
 		}
@@ -18,7 +18,7 @@ __global__ void naive_gemm(double* A, double* B, double* C, int M, int K, int N)
 }
 
 
-void dgemm_cuda(double* d_A, double* d_B, double* d_C, int M, int K, int N){
+void dgemm_cuda(float* d_A, float* d_B, float* d_C, int M, int K, int N){
     
 	dim3 block(32, 32);
     dim3 grid((N + 32 - 1) / 32,
